fix(prefab): only use y rotation on movement facing direction
The facing direction offset on the movement was previously using
all axes for the rotation facing direction, which meant if the facing
direction object (e.g. headset) was pointing at the ground then the
facing direction would be straight down and this would mean the
target would not move.

A new feature in Zinnia means the facing direction can determine
which axes want to be used for the rotational offset so this is now
used and only the y rotation is used so it works over a simple 2 plane.