fix(Prefabs): set Time Multiplier property on MovementMutator
The CombinedActions package has now been updated so the AxesToVector3
prefab exposes the Time Component Extractor multiplier on the Facade
so instead of manually overriding the Time Component at the lower
level, it can now be changed on the facade.

This brings in that change, plus renames the prefabs to something
more meaningful.