fix(Prefabs): ensure smooth rotate uses delta time for time multiplier
There was an issue where smooth rotate would increase in speed with
every rotation due to the Time Multiplier being set to `Time` so it
was increasing from the original time frame.

The Rotation Mutator used within the Smooth Rotate operation is now
using the `Time Step Type Delta Time` Time Multiplier so it should
not increase in speed with every rotation.